@@ -8,7 +8,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.0.3 <4.0.0'
+  sdk: '>=3.3.0 <4.0.0'
 
 
 dependencies:
@@ -23,7 +23,7 @@ dependencies:
   injectable: ^2.1.2
 
   # route
-  go_router: ^7.1.1
+  go_router: ^13.2.0
   
   # data class
   freezed_annotation: ^2.2.0
@@ -36,12 +36,12 @@ dependencies:
   hive_flutter: ^1.1.0
 
   # logger
-  logger: ^1.1.0
+  logger: ^2.0.2+1
 
   # util
   collection: ^1.17.1
   rxdart: ^0.27.7
-  infinite_scroll_pagination: ^3.2.0
+  infinite_scroll_pagination: ^4.0.0
 
   cupertino_icons: ^1.0.2
 
@@ -49,7 +49,7 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  flutter_lints: ^2.0.0
+  flutter_lints: ^3.0.1
   bloc_test:
   build_runner: 
   freezed: 
