@@ -1,7 +1,7 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
 part of 'authentication.dart';
 
@@ -12,7 +12,7 @@ part of 'authentication.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 /// @nodoc
 mixin _$Authentication {
@@ -69,22 +69,22 @@ class _$AuthenticationCopyWithImpl<$Res, $Val extends Authentication>
 }
 
 /// @nodoc
-abstract class _$$_AuthenticationCopyWith<$Res>
+abstract class _$$AuthenticationImplCopyWith<$Res>
     implements $AuthenticationCopyWith<$Res> {
-  factory _$$_AuthenticationCopyWith(
-          _$_Authentication value, $Res Function(_$_Authentication) then) =
-      __$$_AuthenticationCopyWithImpl<$Res>;
+  factory _$$AuthenticationImplCopyWith(_$AuthenticationImpl value,
+          $Res Function(_$AuthenticationImpl) then) =
+      __$$AuthenticationImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({String accessToken, String? refreshToken, String? tokenType});
 }
 
 /// @nodoc
-class __$$_AuthenticationCopyWithImpl<$Res>
-    extends _$AuthenticationCopyWithImpl<$Res, _$_Authentication>
-    implements _$$_AuthenticationCopyWith<$Res> {
-  __$$_AuthenticationCopyWithImpl(
-      _$_Authentication _value, $Res Function(_$_Authentication) _then)
+class __$$AuthenticationImplCopyWithImpl<$Res>
+    extends _$AuthenticationCopyWithImpl<$Res, _$AuthenticationImpl>
+    implements _$$AuthenticationImplCopyWith<$Res> {
+  __$$AuthenticationImplCopyWithImpl(
+      _$AuthenticationImpl _value, $Res Function(_$AuthenticationImpl) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -94,7 +94,7 @@ class __$$_AuthenticationCopyWithImpl<$Res>
     Object? refreshToken = freezed,
     Object? tokenType = freezed,
   }) {
-    return _then(_$_Authentication(
+    return _then(_$AuthenticationImpl(
       accessToken: null == accessToken
           ? _value.accessToken
           : accessToken // ignore: cast_nullable_to_non_nullable
@@ -113,8 +113,8 @@ class __$$_AuthenticationCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$_Authentication implements _Authentication {
-  const _$_Authentication(
+class _$AuthenticationImpl implements _Authentication {
+  const _$AuthenticationImpl(
       {required this.accessToken, this.refreshToken, this.tokenType});
 
   @override
@@ -130,10 +130,10 @@ class _$_Authentication implements _Authentication {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_Authentication &&
+            other is _$AuthenticationImpl &&
             (identical(other.accessToken, accessToken) ||
                 other.accessToken == accessToken) &&
             (identical(other.refreshToken, refreshToken) ||
@@ -149,15 +149,16 @@ class _$_Authentication implements _Authentication {
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$_AuthenticationCopyWith<_$_Authentication> get copyWith =>
-      __$$_AuthenticationCopyWithImpl<_$_Authentication>(this, _$identity);
+  _$$AuthenticationImplCopyWith<_$AuthenticationImpl> get copyWith =>
+      __$$AuthenticationImplCopyWithImpl<_$AuthenticationImpl>(
+          this, _$identity);
 }
 
 abstract class _Authentication implements Authentication {
   const factory _Authentication(
       {required final String accessToken,
       final String? refreshToken,
-      final String? tokenType}) = _$_Authentication;
+      final String? tokenType}) = _$AuthenticationImpl;
 
   @override
   String get accessToken;
@@ -167,6 +168,6 @@ abstract class _Authentication implements Authentication {
   String? get tokenType;
   @override
   @JsonKey(ignore: true)
-  _$$_AuthenticationCopyWith<_$_Authentication> get copyWith =>
+  _$$AuthenticationImplCopyWith<_$AuthenticationImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
