@@ -1,7 +1,7 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
 part of 'sign_in_usecase_param.dart';
 
@@ -12,7 +12,7 @@ part of 'sign_in_usecase_param.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 /// @nodoc
 mixin _$SignInUseCaseParam {
@@ -63,22 +63,22 @@ class _$SignInUseCaseParamCopyWithImpl<$Res, $Val extends SignInUseCaseParam>
 }
 
 /// @nodoc
-abstract class _$$_SignInUseCaseParamCopyWith<$Res>
+abstract class _$$SignInUseCaseParamImplCopyWith<$Res>
     implements $SignInUseCaseParamCopyWith<$Res> {
-  factory _$$_SignInUseCaseParamCopyWith(_$_SignInUseCaseParam value,
-          $Res Function(_$_SignInUseCaseParam) then) =
-      __$$_SignInUseCaseParamCopyWithImpl<$Res>;
+  factory _$$SignInUseCaseParamImplCopyWith(_$SignInUseCaseParamImpl value,
+          $Res Function(_$SignInUseCaseParamImpl) then) =
+      __$$SignInUseCaseParamImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({String email, String password});
 }
 
 /// @nodoc
-class __$$_SignInUseCaseParamCopyWithImpl<$Res>
-    extends _$SignInUseCaseParamCopyWithImpl<$Res, _$_SignInUseCaseParam>
-    implements _$$_SignInUseCaseParamCopyWith<$Res> {
-  __$$_SignInUseCaseParamCopyWithImpl(
-      _$_SignInUseCaseParam _value, $Res Function(_$_SignInUseCaseParam) _then)
+class __$$SignInUseCaseParamImplCopyWithImpl<$Res>
+    extends _$SignInUseCaseParamCopyWithImpl<$Res, _$SignInUseCaseParamImpl>
+    implements _$$SignInUseCaseParamImplCopyWith<$Res> {
+  __$$SignInUseCaseParamImplCopyWithImpl(_$SignInUseCaseParamImpl _value,
+      $Res Function(_$SignInUseCaseParamImpl) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -87,7 +87,7 @@ class __$$_SignInUseCaseParamCopyWithImpl<$Res>
     Object? email = null,
     Object? password = null,
   }) {
-    return _then(_$_SignInUseCaseParam(
+    return _then(_$SignInUseCaseParamImpl(
       email: null == email
           ? _value.email
           : email // ignore: cast_nullable_to_non_nullable
@@ -102,8 +102,8 @@ class __$$_SignInUseCaseParamCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$_SignInUseCaseParam implements _SignInUseCaseParam {
-  const _$_SignInUseCaseParam({required this.email, required this.password});
+class _$SignInUseCaseParamImpl implements _SignInUseCaseParam {
+  const _$SignInUseCaseParamImpl({required this.email, required this.password});
 
   @override
   final String email;
@@ -116,10 +116,10 @@ class _$_SignInUseCaseParam implements _SignInUseCaseParam {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$_SignInUseCaseParam &&
+            other is _$SignInUseCaseParamImpl &&
             (identical(other.email, email) || other.email == email) &&
             (identical(other.password, password) ||
                 other.password == password));
@@ -131,15 +131,15 @@ class _$_SignInUseCaseParam implements _SignInUseCaseParam {
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$_SignInUseCaseParamCopyWith<_$_SignInUseCaseParam> get copyWith =>
-      __$$_SignInUseCaseParamCopyWithImpl<_$_SignInUseCaseParam>(
+  _$$SignInUseCaseParamImplCopyWith<_$SignInUseCaseParamImpl> get copyWith =>
+      __$$SignInUseCaseParamImplCopyWithImpl<_$SignInUseCaseParamImpl>(
           this, _$identity);
 }
 
 abstract class _SignInUseCaseParam implements SignInUseCaseParam {
   const factory _SignInUseCaseParam(
       {required final String email,
-      required final String password}) = _$_SignInUseCaseParam;
+      required final String password}) = _$SignInUseCaseParamImpl;
 
   @override
   String get email;
@@ -147,6 +147,6 @@ abstract class _SignInUseCaseParam implements SignInUseCaseParam {
   String get password;
   @override
   @JsonKey(ignore: true)
-  _$$_SignInUseCaseParamCopyWith<_$_SignInUseCaseParam> get copyWith =>
+  _$$SignInUseCaseParamImplCopyWith<_$SignInUseCaseParamImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
