@@ -1,7 +1,7 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
 part of 'sign_in_bloc.dart';
 
@@ -12,7 +12,7 @@ part of 'sign_in_bloc.dart';
 T _$identity<T>(T value) => value;
 
 final _privateConstructorUsedError = UnsupportedError(
-    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
 
 /// @nodoc
 mixin _$SignInEvent {
@@ -95,22 +95,22 @@ class _$SignInEventCopyWithImpl<$Res, $Val extends SignInEvent>
 }
 
 /// @nodoc
-abstract class _$$SignInRequestedCopyWith<$Res>
+abstract class _$$SignInRequestedImplCopyWith<$Res>
     implements $SignInEventCopyWith<$Res> {
-  factory _$$SignInRequestedCopyWith(
-          _$SignInRequested value, $Res Function(_$SignInRequested) then) =
-      __$$SignInRequestedCopyWithImpl<$Res>;
+  factory _$$SignInRequestedImplCopyWith(_$SignInRequestedImpl value,
+          $Res Function(_$SignInRequestedImpl) then) =
+      __$$SignInRequestedImplCopyWithImpl<$Res>;
   @override
   @useResult
   $Res call({String email, String password});
 }
 
 /// @nodoc
-class __$$SignInRequestedCopyWithImpl<$Res>
-    extends _$SignInEventCopyWithImpl<$Res, _$SignInRequested>
-    implements _$$SignInRequestedCopyWith<$Res> {
-  __$$SignInRequestedCopyWithImpl(
-      _$SignInRequested _value, $Res Function(_$SignInRequested) _then)
+class __$$SignInRequestedImplCopyWithImpl<$Res>
+    extends _$SignInEventCopyWithImpl<$Res, _$SignInRequestedImpl>
+    implements _$$SignInRequestedImplCopyWith<$Res> {
+  __$$SignInRequestedImplCopyWithImpl(
+      _$SignInRequestedImpl _value, $Res Function(_$SignInRequestedImpl) _then)
       : super(_value, _then);
 
   @pragma('vm:prefer-inline')
@@ -119,7 +119,7 @@ class __$$SignInRequestedCopyWithImpl<$Res>
     Object? email = null,
     Object? password = null,
   }) {
-    return _then(_$SignInRequested(
+    return _then(_$SignInRequestedImpl(
       email: null == email
           ? _value.email
           : email // ignore: cast_nullable_to_non_nullable
@@ -134,8 +134,8 @@ class __$$SignInRequestedCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$SignInRequested implements SignInRequested {
-  const _$SignInRequested({required this.email, required this.password});
+class _$SignInRequestedImpl implements SignInRequested {
+  const _$SignInRequestedImpl({required this.email, required this.password});
 
   @override
   final String email;
@@ -148,10 +148,10 @@ class _$SignInRequested implements SignInRequested {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _$SignInRequested &&
+            other is _$SignInRequestedImpl &&
             (identical(other.email, email) || other.email == email) &&
             (identical(other.password, password) ||
                 other.password == password));
@@ -163,8 +163,9 @@ class _$SignInRequested implements SignInRequested {
   @JsonKey(ignore: true)
   @override
   @pragma('vm:prefer-inline')
-  _$$SignInRequestedCopyWith<_$SignInRequested> get copyWith =>
-      __$$SignInRequestedCopyWithImpl<_$SignInRequested>(this, _$identity);
+  _$$SignInRequestedImplCopyWith<_$SignInRequestedImpl> get copyWith =>
+      __$$SignInRequestedImplCopyWithImpl<_$SignInRequestedImpl>(
+          this, _$identity);
 
   @override
   @optionalTypeArgs
@@ -226,7 +227,7 @@ class _$SignInRequested implements SignInRequested {
 abstract class SignInRequested implements SignInEvent {
   const factory SignInRequested(
       {required final String email,
-      required final String password}) = _$SignInRequested;
+      required final String password}) = _$SignInRequestedImpl;
 
   @override
   String get email;
@@ -234,7 +235,7 @@ abstract class SignInRequested implements SignInEvent {
   String get password;
   @override
   @JsonKey(ignore: true)
-  _$$SignInRequestedCopyWith<_$SignInRequested> get copyWith =>
+  _$$SignInRequestedImplCopyWith<_$SignInRequestedImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
 
@@ -311,25 +312,25 @@ class _$SignInStateCopyWithImpl<$Res, $Val extends SignInState>
 }
 
 /// @nodoc
-abstract class _$$SignInInitialCopyWith<$Res> {
-  factory _$$SignInInitialCopyWith(
-          _$SignInInitial value, $Res Function(_$SignInInitial) then) =
-      __$$SignInInitialCopyWithImpl<$Res>;
+abstract class _$$SignInInitialImplCopyWith<$Res> {
+  factory _$$SignInInitialImplCopyWith(
+          _$SignInInitialImpl value, $Res Function(_$SignInInitialImpl) then) =
+      __$$SignInInitialImplCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$$SignInInitialCopyWithImpl<$Res>
-    extends _$SignInStateCopyWithImpl<$Res, _$SignInInitial>
-    implements _$$SignInInitialCopyWith<$Res> {
-  __$$SignInInitialCopyWithImpl(
-      _$SignInInitial _value, $Res Function(_$SignInInitial) _then)
+class __$$SignInInitialImplCopyWithImpl<$Res>
+    extends _$SignInStateCopyWithImpl<$Res, _$SignInInitialImpl>
+    implements _$$SignInInitialImplCopyWith<$Res> {
+  __$$SignInInitialImplCopyWithImpl(
+      _$SignInInitialImpl _value, $Res Function(_$SignInInitialImpl) _then)
       : super(_value, _then);
 }
 
 /// @nodoc
 
-class _$SignInInitial implements SignInInitial {
-  const _$SignInInitial();
+class _$SignInInitialImpl implements SignInInitial {
+  const _$SignInInitialImpl();
 
   @override
   String toString() {
@@ -337,9 +338,9 @@ class _$SignInInitial implements SignInInitial {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$SignInInitial);
+        (other.runtimeType == runtimeType && other is _$SignInInitialImpl);
   }
 
   @override
@@ -421,29 +422,29 @@ class _$SignInInitial implements SignInInitial {
 }
 
 abstract class SignInInitial implements SignInState {
-  const factory SignInInitial() = _$SignInInitial;
+  const factory SignInInitial() = _$SignInInitialImpl;
 }
 
 /// @nodoc
-abstract class _$$SignInLoadingCopyWith<$Res> {
-  factory _$$SignInLoadingCopyWith(
-          _$SignInLoading value, $Res Function(_$SignInLoading) then) =
-      __$$SignInLoadingCopyWithImpl<$Res>;
+abstract class _$$SignInLoadingImplCopyWith<$Res> {
+  factory _$$SignInLoadingImplCopyWith(
+          _$SignInLoadingImpl value, $Res Function(_$SignInLoadingImpl) then) =
+      __$$SignInLoadingImplCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$$SignInLoadingCopyWithImpl<$Res>
-    extends _$SignInStateCopyWithImpl<$Res, _$SignInLoading>
-    implements _$$SignInLoadingCopyWith<$Res> {
-  __$$SignInLoadingCopyWithImpl(
-      _$SignInLoading _value, $Res Function(_$SignInLoading) _then)
+class __$$SignInLoadingImplCopyWithImpl<$Res>
+    extends _$SignInStateCopyWithImpl<$Res, _$SignInLoadingImpl>
+    implements _$$SignInLoadingImplCopyWith<$Res> {
+  __$$SignInLoadingImplCopyWithImpl(
+      _$SignInLoadingImpl _value, $Res Function(_$SignInLoadingImpl) _then)
       : super(_value, _then);
 }
 
 /// @nodoc
 
-class _$SignInLoading implements SignInLoading {
-  const _$SignInLoading();
+class _$SignInLoadingImpl implements SignInLoading {
+  const _$SignInLoadingImpl();
 
   @override
   String toString() {
@@ -451,9 +452,9 @@ class _$SignInLoading implements SignInLoading {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$SignInLoading);
+        (other.runtimeType == runtimeType && other is _$SignInLoadingImpl);
   }
 
   @override
@@ -535,29 +536,29 @@ class _$SignInLoading implements SignInLoading {
 }
 
 abstract class SignInLoading implements SignInState {
-  const factory SignInLoading() = _$SignInLoading;
+  const factory SignInLoading() = _$SignInLoadingImpl;
 }
 
 /// @nodoc
-abstract class _$$SignInSucceedCopyWith<$Res> {
-  factory _$$SignInSucceedCopyWith(
-          _$SignInSucceed value, $Res Function(_$SignInSucceed) then) =
-      __$$SignInSucceedCopyWithImpl<$Res>;
+abstract class _$$SignInSucceedImplCopyWith<$Res> {
+  factory _$$SignInSucceedImplCopyWith(
+          _$SignInSucceedImpl value, $Res Function(_$SignInSucceedImpl) then) =
+      __$$SignInSucceedImplCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$$SignInSucceedCopyWithImpl<$Res>
-    extends _$SignInStateCopyWithImpl<$Res, _$SignInSucceed>
-    implements _$$SignInSucceedCopyWith<$Res> {
-  __$$SignInSucceedCopyWithImpl(
-      _$SignInSucceed _value, $Res Function(_$SignInSucceed) _then)
+class __$$SignInSucceedImplCopyWithImpl<$Res>
+    extends _$SignInStateCopyWithImpl<$Res, _$SignInSucceedImpl>
+    implements _$$SignInSucceedImplCopyWith<$Res> {
+  __$$SignInSucceedImplCopyWithImpl(
+      _$SignInSucceedImpl _value, $Res Function(_$SignInSucceedImpl) _then)
       : super(_value, _then);
 }
 
 /// @nodoc
 
-class _$SignInSucceed implements SignInSucceed {
-  const _$SignInSucceed();
+class _$SignInSucceedImpl implements SignInSucceed {
+  const _$SignInSucceedImpl();
 
   @override
   String toString() {
@@ -565,9 +566,9 @@ class _$SignInSucceed implements SignInSucceed {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$SignInSucceed);
+        (other.runtimeType == runtimeType && other is _$SignInSucceedImpl);
   }
 
   @override
@@ -649,29 +650,29 @@ class _$SignInSucceed implements SignInSucceed {
 }
 
 abstract class SignInSucceed implements SignInState {
-  const factory SignInSucceed() = _$SignInSucceed;
+  const factory SignInSucceed() = _$SignInSucceedImpl;
 }
 
 /// @nodoc
-abstract class _$$SignInErrorCopyWith<$Res> {
-  factory _$$SignInErrorCopyWith(
-          _$SignInError value, $Res Function(_$SignInError) then) =
-      __$$SignInErrorCopyWithImpl<$Res>;
+abstract class _$$SignInErrorImplCopyWith<$Res> {
+  factory _$$SignInErrorImplCopyWith(
+          _$SignInErrorImpl value, $Res Function(_$SignInErrorImpl) then) =
+      __$$SignInErrorImplCopyWithImpl<$Res>;
 }
 
 /// @nodoc
-class __$$SignInErrorCopyWithImpl<$Res>
-    extends _$SignInStateCopyWithImpl<$Res, _$SignInError>
-    implements _$$SignInErrorCopyWith<$Res> {
-  __$$SignInErrorCopyWithImpl(
-      _$SignInError _value, $Res Function(_$SignInError) _then)
+class __$$SignInErrorImplCopyWithImpl<$Res>
+    extends _$SignInStateCopyWithImpl<$Res, _$SignInErrorImpl>
+    implements _$$SignInErrorImplCopyWith<$Res> {
+  __$$SignInErrorImplCopyWithImpl(
+      _$SignInErrorImpl _value, $Res Function(_$SignInErrorImpl) _then)
       : super(_value, _then);
 }
 
 /// @nodoc
 
-class _$SignInError implements SignInError {
-  const _$SignInError();
+class _$SignInErrorImpl implements SignInError {
+  const _$SignInErrorImpl();
 
   @override
   String toString() {
@@ -679,9 +680,9 @@ class _$SignInError implements SignInError {
   }
 
   @override
-  bool operator ==(dynamic other) {
+  bool operator ==(Object other) {
     return identical(this, other) ||
-        (other.runtimeType == runtimeType && other is _$SignInError);
+        (other.runtimeType == runtimeType && other is _$SignInErrorImpl);
   }
 
   @override
@@ -763,5 +764,5 @@ class _$SignInError implements SignInError {
 }
 
 abstract class SignInError implements SignInState {
-  const factory SignInError() = _$SignInError;
+  const factory SignInError() = _$SignInErrorImpl;
 }
