@@ -1,4 +1,3 @@
-import 'package:bloc_architecture/feature/common/presentation/main_bottom_navigation_bar.dart';
 import 'package:flutter/material.dart';
 
 class HomePage extends StatelessWidget {
@@ -6,12 +5,13 @@ class HomePage extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      bottomNavigationBar: const MainBottomNavigationBar(),
-      appBar: AppBar(
-        automaticallyImplyLeading: false,
-        title: const Text('HomePage'),
-      ),
+    return Column(
+      children: [
+        AppBar(
+          automaticallyImplyLeading: false,
+          title: const Text('HomePage'),
+        ),
+      ],
     );
   }
 }
