@@ -1,8 +1,10 @@
+import 'package:bloc_architecture/feature/auth/domain/usecase/sign_in_usecase.dart';
 import 'package:bloc_architecture/feature/auth/presentation/sign_in/bloc/sign_in_bloc.dart';
 import 'package:bloc_architecture/feature/auth/presentation/sign_in/sign_in_page.dart';
 import 'package:bloc_architecture/feature/chat/presentation/chat_page.dart';
 import 'package:bloc_architecture/feature/community/presentation/community_page.dart';
 import 'package:bloc_architecture/feature/home/presentation/home_page.dart';
+import 'package:bloc_architecture/feature/main/presentation/main_screen.dart';
 import 'package:bloc_architecture/feature/user/presentation/profile_page.dart';
 import 'package:bloc_architecture/injection/injection.dart';
 import 'package:flutter/material.dart';
@@ -22,56 +24,51 @@ final routes = GoRouter(
     GoRoute(
       path: '/${AppRouteState.login.path}',
       name: AppRouteState.login.name,
-      pageBuilder: (context, state) => appPageBuilder(
-        BlocProvider(
-          create: (context) => SignInBloc(signInUseCase: getIt()),
-          child: const SignInPage(),
-        ),
+      builder: (context, state) => BlocProvider(
+        create: (context) => SignInBloc(signInUseCase: getIt<SignInUseCase>()),
+        child: const SignInPage(),
       ),
     ),
-    ShellRoute(
-      navigatorKey: shellRouteNavigatorKey,
-      pageBuilder: (context, state, child) => appPageBuilder(child),
-      routes: [
-        GoRoute(
-          path: '/${AppRouteState.home.path}',
-          name: AppRouteState.home.name,
-          pageBuilder: (context, state) => appPageBuilder(
-            const HomePage(),
-          ),
+    StatefulShellRoute.indexedStack(
+      builder: (context, state, navigationShell) => MainScreen(navigationShell: navigationShell),
+      branches: [
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/${AppRouteState.home.path}',
+              name: AppRouteState.home.name,
+              builder: (context, state) => const HomePage(),
+            ),
+          ],
         ),
-        GoRoute(
-          path: '/${AppRouteState.community.path}',
-          name: AppRouteState.community.name,
-          pageBuilder: (context, state) => appPageBuilder(
-            const CommunityPage(),
-          ),
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/${AppRouteState.community.path}',
+              name: AppRouteState.community.name,
+              builder: (context, state) => const CommunityPage(),
+            ),
+          ],
         ),
-        GoRoute(
-          path: '/${AppRouteState.chat.path}',
-          name: AppRouteState.chat.name,
-          pageBuilder: (context, state) => appPageBuilder(
-            const ChatPage(),
-          ),
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/${AppRouteState.chat.path}',
+              name: AppRouteState.chat.name,
+              builder: (context, state) => const ChatPage(),
+            ),
+          ],
         ),
-        GoRoute(
-          path: '/${AppRouteState.profile.path}',
-          name: AppRouteState.profile.name,
-          pageBuilder: (context, state) => appPageBuilder(
-            const ProfilePage(),
-          ),
+        StatefulShellBranch(
+          routes: [
+            GoRoute(
+              path: '/${AppRouteState.profile.path}',
+              name: AppRouteState.profile.name,
+              builder: (context, state) => const ProfilePage(),
+            ),
+          ],
         ),
       ],
     ),
   ],
 );
-
-Page<dynamic> appPageBuilder(Widget child, {bool isNoTransition = true}) =>
-    CustomTransitionPage(
-      child: child,
-      transitionDuration: const Duration(milliseconds: 200),
-      transitionsBuilder: (context, animation, secondAnimation, child) =>
-          isNoTransition
-              ? child
-              : FadeTransition(opacity: animation, child: child),
-    );
